requirements.txt
absl-py==1.0.0
ale-py==0.7.4
black==22.1.0
box2d-py==2.3.5
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
Cython==0.29.28
docker-pycreds==0.4.0
fasteners==0.17.3
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.23.1
gym-notices==0.0.6
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.4.0
lockfile==0.12.2
lz4==4.0.0
Markdown==3.3.6
mypy-extensions==0.4.3
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
pandas==1.3.5
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.0.1
platformdirs==2.5.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
sentry-sdk==1.5.7
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tomli==2.0.1
torch
torch-tb-profiler==0.4.0
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.9
wandb==0.12.11
Werkzeug==2.1.1
yaspin==2.1.0
zipp==3.7.0
